feat: Complete backend deployment readiness with AI features and comprehensive endpoints
- Add Redis caching system with configurable TTLs
- Implement tiered rate limiting (public/auth/AI/admin tiers)
- Add input validation and sanitization with Pydantic models
- Create custom middleware (error handling, security headers, request ID, timing)
- Add 41+ API endpoints with full CRUD operations
- Implement database repositories (workout sessions, notifications, progress tracking)
- Add AI memory/context persistence
- Create database migration system
- Add comprehensive test suite (90+ tests)
- Setup CI/CD pipeline with GitHub Actions
- Create deployment documentation
- Integrate caching and rate limiting into main app
- Add logging configuration with Sentry support
- Create validation utilities for all input types

Test Plan: Run python validate_backend.py && python run_tests.py

+++ app/lib/models/workout_tracking.dart
@@ -290,6 +290,6 @@ class MicroCycle {
   }
 
   List<double> get rpeTrend {
-    return days.expand((d) => d.rpeEntries).toList();
+    return days.expand((d) => d.rpeEntries).cast<double>().toList();
   }
 }
